Add files via upload
I learned web scraping through an eDX course for data science. When asked to do a prospecting list for F&B deliveries, I thought of using it to speed up my process by using BeautifulSoup. While I was unable to retrieve the phone numbers and email address because I was unfamiliar with AJAX and Postman, I had managed to sieve out the restaurant names and cuisine type, which did not pop-up in-site. This helped me to expedite my process of creating a list to begin with.